Split inference prior weights into laplace, asset, and catalog modules.
Decompose the monolithic loader so parsing, Laplace conversion, and catalog
resolution live in focused modules with a shared ship-limit band parser.
